+++ dtool/src/interrogate/functionRemap.cxx
@@ -767,12 +767,12 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
   }
 
   // Check for a special meaning by name and signature.
-  int first_param = 0;
+  size_t first_param = 0;
   if (_has_this) {
     first_param = 1;
   }
 
-  if (_parameters.size() > (size_t)first_param && _parameters[first_param]._name == "self" &&
+  if (_parameters.size() > first_param && _parameters[first_param]._name == "self" &&
       TypeManager::is_pointer_to_PyObject(_parameters[first_param]._remap->get_orig_type())) {
     // Here's a special case.  If the first parameter of a nonstatic method
     // is a PyObject * called "self", then we will automatically fill it in
@@ -782,9 +782,9 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
     _flags |= F_explicit_self;
   }
 
-  if ((int)_parameters.size() == first_param) {
+  if (_parameters.size() == first_param) {
     _args_type = InterfaceMaker::AT_no_args;
-  } else if ((int)_parameters.size() == first_param + 1 &&
+  } else if (_parameters.size() == first_param + 1 &&
              _parameters[first_param]._remap->get_default_value() == NULL) {
     _args_type = InterfaceMaker::AT_single_arg;
   } else {
@@ -833,7 +833,7 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
       }
 
     } else if (fname == "size" || fname == "__len__") {
-      if ((int)_parameters.size() == first_param &&
+      if (_parameters.size() == first_param &&
           TypeManager::is_integer(_return_type->get_new_type())) {
         // It receives no parameters, and returns an integer.
         _flags |= F_size;
@@ -847,7 +847,7 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
       }
 
     } else if (fname == "__iter__") {
-      if ((int)_parameters.size() == first_param &&
+      if (_parameters.size() == first_param &&
           TypeManager::is_pointer(_return_type->get_new_type())) {
         // It receives no parameters, and returns a pointer.
         _flags |= F_iter;
@@ -870,7 +870,7 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
       if (_args_type == InterfaceMaker::AT_varargs) {
         // Of course methods named "make" can still take kwargs, if they are
         // named.
-        for (int i = first_param; i < _parameters.size(); ++i) {
+        for (size_t i = first_param; i < _parameters.size(); ++i) {
           if (_parameters[i]._has_name) {
             _args_type = InterfaceMaker::AT_keyword_args;
             break;
@@ -904,7 +904,7 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
       if (_args_type == InterfaceMaker::AT_varargs) {
         // Every other method can take keyword arguments, if they take more
         // than one argument, and the arguments are named.
-        for (int i = first_param; i < _parameters.size(); ++i) {
+        for (size_t i = first_param; i < _parameters.size(); ++i) {
           if (_parameters[i]._has_name) {
             _args_type |= InterfaceMaker::AT_keyword_args;
             break;
@@ -960,7 +960,7 @@ setup_properties(const InterrogateFunction &ifunc, InterfaceMaker *interface_mak
 
     // Constructors always take varargs, and possibly keyword args.
     _args_type = InterfaceMaker::AT_varargs;
-    for (int i = first_param; i < _parameters.size(); ++i) {
+    for (size_t i = first_param; i < _parameters.size(); ++i) {
       if (_parameters[i]._has_name) {
         _args_type = InterfaceMaker::AT_keyword_args;
         break;
